Add files via upload
These are template scripts which I used Monday through Thursday. This could all be automated in 1 script and linked to a time function which would recognise which day sequence it would run. This one script could then be connected to the Windows task scheduler but I couldn't make it work at that time.